+++ src/components/App.jsx
@@ -87,11 +87,10 @@ class App extends Component {
           <Typography paragraph>This tool was made in React using the GitHub API to fetch the languages used in your or anyone's GitHub repositories, and display them including how much of the repository
             is made of that language.
           </Typography>
-          <Typography paragraph>You can use this tool on your network up to 60 times per hour if you do not provide a personal access token. If you do want to use this more than 60 times an hour, you must
-            create your own personal access token. To do this, you have to create a GitHub account if you haven't already, and go
-            to <a href='https://github.com/settings/tokens' target='_blank' rel='noreferrer noopener'>https://github.com/settings/tokens</a>. Click on "Generate new token", type a small description like
-            "GitHub Repo Langs", check the repo checkbox, and click the generate token button at the bottom. Copy the token, <strong>KEEP THIS TOKEN SECRET</strong>, and paste it in the text box
-            below. Doing this allows you to increase the amount of times you can use this, and you can get to see the languages on your private repositories.
+          <Typography paragraph>This app currently does not support private repositories, if you want to see your own private repositories, you must create your own personal access token. To do this,
+            you have to create a GitHub account if you haven't already, and go to <a href='https://github.com/settings/tokens' target='_blank' rel='noreferrer noopener'>https://github.com/settings/tokens</a>.
+            Click on "Generate new token", type a small description like "GitHub Repo Langs", check the repo checkbox, and click the generate token button at the bottom. Copy the token, <strong>KEEP THIS TOKEN
+            SECRET</strong>, and paste it in the text box below.
           </Typography>
           <Typography paragraph>If nothing shows up when you click the go button, one of three things must of happened. Either you mistyped the username, repository (or it doesn't exist), token (or it's invalid)
             or you hit the rate limit for the GitHub API and must follow the instructions above, or the repository is private.